+++ example/fastmath/interpolation_examples.clj
@@ -3,7 +3,7 @@
             [incanter.charts :as c]
             [fastmath.core :as m]
             [metadoc.examples :refer :all]
-            [fastmath.rbf :as rbf]
+            [fastmath.kernel :as k]
             [incanter.core :as i]))
 
 (defsnippet fastmath.interpolation save-graph
@@ -87,16 +87,17 @@
        (intrp 5 6)])))
 
 (add-examples rbf
-  (example-snippet "RBF - Linear" save-interpolation :image (fn [xs ys] (rbf (rbf/rbf :linear) xs ys)) 0 7)
-  (example-snippet "RBF - Gaussian" save-interpolation :image (fn [xs ys] (rbf (rbf/rbf :gaussian) xs ys)) 0 7)
-  (example-snippet "RBF - Multiquadratic" save-interpolation :image (fn [xs ys] (rbf (rbf/rbf :multiquadratic) xs ys)) 0 7)
-  (example-snippet "RBF - Inverse ultiquadratic" save-interpolation :image (fn [xs ys] (rbf (rbf/rbf :inverse-multiquadratic) xs ys)) 0 7)
-  (example-snippet "RBF - Inverse quadratic" save-interpolation :image (fn [xs ys] (rbf (rbf/rbf :inverse-quadratic) xs ys)) 0 7)
-  (example-snippet "RBF - Thinplate" save-interpolation :image (fn [xs ys] (rbf (rbf/rbf :thinplate) xs ys)) 0 7)
-  (example-snippet "RBF - Polyharmonic 3" save-interpolation :image (fn [xs ys] (rbf (rbf/rbf :polyharmonic 3) xs ys)) 0 7)
-  (example-snippet "RBF - Polyharmonic 4" save-interpolation :image (fn [xs ys] (rbf (rbf/rbf :polyharmonic 4) xs ys)) 0 7)
-  (example-snippet "RBF - Wendland" save-interpolation :image (fn [xs ys] (rbf (rbf/rbf :wendland) xs ys)) 0 7)
-  (example-snippet "RBF - Wu" save-interpolation :image (fn [xs ys] (rbf (rbf/rbf :wu) xs ys)) 0 7))
+              (example-snippet "RBF - Linear" save-interpolation :image (fn [xs ys] (rbf (k/rbf :linear) xs ys)) 0 7)
+              (example-snippet "RBF - Gaussian" save-interpolation :image (fn [xs ys] (rbf (k/rbf :gaussian) xs ys)) 0 7)
+              (example-snippet "RBF - Multiquadratic" save-interpolation :image (fn [xs ys] (rbf (k/rbf :multiquadratic) xs ys)) 0 7)
+              (example-snippet "RBF - Inverse ultiquadratic" save-interpolation :image (fn [xs ys] (rbf (k/rbf :inverse-multiquadratic) xs ys)) 0 7)
+              ;; (example-snippet "RBF - Inverse quadratic" save-interpolation :image (fn [xs ys] (rbf (k/rbf :inverse-quadratic) xs ys)) 0 7)
+              (example-snippet "RBF - Thinplate" save-interpolation :image (fn [xs ys] (rbf (k/rbf :thinplate) xs ys)) 0 7)
+              ;; (example-snippet "RBF - Polyharmonic 3" save-interpolation :image (fn [xs ys] (rbf (k/rbf :polyharmonic 3) xs ys)) 0 7)
+              ;; (example-snippet "RBF - Polyharmonic 4" save-interpolation :image (fn [xs ys] (rbf (k/rbf :polyharmonic 4) xs ys)) 0 7)
+              ;; (example-snippet "RBF - Wendland" save-interpolation :image (fn [xs ys] (rbf (k/rbf :wendland) xs ys)) 0 7)
+              ;; (example-snippet "RBF - Wu" save-interpolation :image (fn [xs ys] (rbf (k/rbf :wu) xs ys)) 0 7)
+              )
 
 
 (add-examples interpolators-1d-list
@@ -105,7 +106,7 @@
 (add-examples interpolators-2d-list
   (example "List of names" (keys interpolators-2d-list)))
 
-(comment let [f (fn [xs ys] (rbf xs ys (rbf/rbf :wu 2)))
+(comment let [f (fn [xs ys] (rbf xs ys (k/rbf :wu 2)))
               xs [0.69 1.73 2.0 2.28 3.46 4.18 4.84 5.18 5.53 5.87 6.22]
               ft (fn [^double x] (m/sin (* x (* 0.5 (m/cos (inc x))))))
               ys (map ft xs)